backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "aiofiles>=24.1.0",
    "email-validator>=2.2.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.12",
    "alembic>=1.14.0",
    "python-multipart>=0.0.7",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "pydantic-settings>=2.0",
    "loguru>=0.7.2",
    "grpcio>=1.68.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.23.6",
    "ruff==0.11.13",
    "types-aiofiles>=23.1.0.5",
    "grpcio-tools>=1.68.0",
    "types-grpcio>=1.0.0.20250525",
]

[tool.ruff]
line-length = 100
required-version = "0.11.13"
target-version = "py310"
extend-exclude = ["migrations", "scripts"]

[tool.ruff.lint]
dummy-variable-rgx = "^_$"
select = ["ALL"]
ignore = [
    "D105",
    # D4..
    "D401",
    "D403",
    "D404",

    # others
    "S311",
    "FBT",
    "G",
    # TD
    "TD001",
    "TD002",
    "TD003",
    # FIX, ERA
    "FIX",
    "ERA",
    # TRY
    "TRY400",
    "TRY401",
    # RUF
    "RUF001",
    "RUF002",
    "RUF003",
    # init
    "D107",
    "ANN204",
    # formatter incompatible
    "COM812",
    "ISC001",
    # sql
    "S608",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
line-ending = "lf"
quote-style = "single"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Body",
    "fastapi.Depends",
    "fastapi.Form",
    "fastapi.File",
    "fastapi.Header",
    "fastapi.Path",
    "fastapi.Query",
    "fastapi.Security",
]

[tool.mypy]
python_version = "3.12"
exclude = ["^migrations/$", "^venv/$"]
plugins = ["pydantic.mypy"]

[tool.setuptools.packages.find]
exclude = ["data*"]

